# backend/openapi.yaml
openapi: 3.0.0
paths:
...
  /shanda/requests:
    get:
      operationId: RequestsController_findAll
      parameters:
        - name: page
          required: false
          in: query
          description: Page number
          schema:
            minimum: 1
            default: 1
            example: 1
            type: number
        - name: limit
          required: false
          in: query
          description: Items per page
          schema:
            minimum: 1
            maximum: 100
            default: 20
            example: 20
            type: number
        - name: status
          required: false
          in: query
          description: Filter by status (open, quoted, in_progress, completed, cancelled)
          schema:
            type: string
        - name: sort
          required: false
          in: query
          description: Sort order
          schema:
            default: newest
            type: string
            enum:
              - newest
              - oldest
              - urgency
      responses:
        '200':
          description: Paginated list of service requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RequestResponseDto'
      security:
        - bearer: []
      summary: Get all service requests (owners see theirs, providers see all open)
      tags: *ref_3
    post:
      operationId: RequestsController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRequestDto'
      responses:
        '201':
          description: Service request created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequestResponseDto'
      security:
        - bearer: []
      summary: Create a new service request
      tags: *ref_3
  /shanda/requests/{id}:
    get:
      operationId: RequestsController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Service request details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequestResponseDto'
        '404':
          description: Request not found.
      security:
        - bearer: []
      summary: Get a specific service request by ID
      tags: *ref_3
    put:
      operationId: RequestsController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRequestDto'
      responses:
        '200':
          description: Service request updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequestResponseDto'
        '404':
          description: Request not found.
      security:
        - bearer: []
      summary: Update a service request
      tags: *ref_3
  /shanda/requests/{id}/images:
    post:
      operationId: RequestsController_uploadImages
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Images uploaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequestResponseDto'
        '201':
          description: ''
      security:
        - bearer: []
      summary: Upload images for a service request
      tags: *ref_3
    delete:
      operationId: RequestsController_deleteImage
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Image deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequestResponseDto'
      security:
        - bearer: []
      summary: Delete an image from a service request
      tags: *ref_3
  /shanda/quotes/request/{requestId}:
    get:
      operationId: QuotesController_findByRequest
      parameters:
        - name: requestId
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/QuoteEntity'
      security: &ref_4
        - bearer: []
      summary: Get all quotes for a service request
      tags: &ref_5
        - quotes
...
info:
  title: Shanda Automotive API
  description: Marketplace connecting vehicle owners with service providers
  version: '1.0'
  contact: {}
...
servers:
  - url: http://localhost:4201
    description: Local Development
  - url: https://api.shanda.com
    description: Production
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    RequestResponseDto:
      type: object
      properties:
        id:
          type: string
          example: 123e4567-e89b-12d3-a456-426614174000
        ownerId:
          type: string
          example: 123e4567-e89b-12d3-a456-426614174001
        vehicleId:
          type: string
          example: 123e4567-e89b-12d3-a456-426614174002
        serviceType:
          type: string
          example: oil_change
        description:
          type: string
          example: Need oil change and filter replacement
        urgency:
          type: string
          enum:
            - low
            - medium
            - high
            - urgent
          example: medium
        preferredLocation:
          type: string
          nullable: true
          example: 123 Main St, City, State 12345
        preferredDate:
          format: date-time
          type: string
          nullable: true
        status:
          enum:
            - open
            - quoted
            - accepted
            - in_progress
            - completed
            - cancelled
          type: string
          example: open
        imageUrls:
          example:
            - https://example.com/image1.jpg
          type: array
          items:
            type: string
        createdAt:
          format: date-time
          type: string
        updatedAt:
          format: date-time
          type: string
      required:
        - id
        - ownerId
        - vehicleId
        - serviceType
        - description
        - urgency
        - status
        - createdAt
        - updatedAt
    CreateRequestDto:
      type: object
      properties:
        vehicleId:
          type: string
          format: uuid
          description: ID of the vehicle (optional - can create inline)
        make:
          type: string
          description: Vehicle make (required if vehicleId not provided)
        model:
          type: string
          description: Vehicle model (required if vehicleId not provided)
        year:
          type: number
          minimum: 1900
          maximum: 2027
          description: Vehicle year (required if vehicleId not provided)
        title:
          type: string
          description: Title of the request
        description:
          type: string
          description: Detailed description of the issue
        urgency:
          enum: &ref_19
            - low
            - medium
            - high
            - urgent
          type: string
          description: Urgency level
      required:
        - title
        - description
        - urgency
    UpdateRequestDto:
      type: object
      properties:
        vehicleId:
          type: string
          format: uuid
          description: ID of the vehicle (optional - can create inline)
        make:
          type: string
          description: Vehicle make (required if vehicleId not provided)
        model:
          type: string
          description: Vehicle model (required if vehicleId not provided)
        year:
          type: number
          minimum: 1900
          maximum: 2027
          description: Vehicle year (required if vehicleId not provided)
        title:
          type: string
          description: Title of the request
        description:
          type: string
          description: Detailed description of the issue
        urgency:
          enum: *ref_19
          type: string
          description: Urgency level
    QuoteEntity:
      type: object
      properties:
        id:
          type: string
        requestId:
          type: string
        providerId:
          type: string
        amount:
          type: string
        laborCost:
          type: string
          nullable: true
        partsCost:
          type: string
          nullable: true
        estimatedDuration:
          type: string
        description:
          type: string
          nullable: true
        includesWarranty:
          type: boolean
        status:
          type: string
        createdAt:
          format: date-time
          type: string
        updatedAt:
          format: date-time
          type: string
        provider:
          type: object
      required:
        - id
        - requestId
        - providerId
        - amount
        - estimatedDuration
        - includesWarranty
        - status
        - createdAt
        - updatedAt
...